Migrate launch file and keep PR2 URDF in pkg tree
The launch file to run the tests has been updated to use ROS 2
constructs.
Moreover, the PR2 URDF file cannot be obtained as a runtime dependency
because `pr2_common` is not released as a ROS 2 package. As such, I have
manually included a `pr2.urdf` file in the package tree.

Signed-off-by: aprotyas <aprotyas@u.rochester.edu>